requirements.txt
asgiref==3.7.2
astroid==2.4.2
colorama==0.4.6
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
Faker==19.2.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
python-dateutil==2.8.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typing_extensions==4.7.1
tzdata==2023.3
validate-docbr==1.10.0
wrapt==1.12.1